+++ src/lib/mrvUI/mrvMonitor_linux.cpp
@@ -16,19 +16,19 @@
 
 namespace mrv
 {
-    namespace desktop
+    namespace monitor
     {
         // Get the monitor name given its FLTK screen index
-        std::string getMonitorName(int monitorIndex)
+        std::string getName(int monitorIndex)
         {
             std::string out;
 #ifdef FLTK_USE_X11
             if (fl_x11_display())
-                out = getX11MonitorName(monitorIndex);
+                out = getX11Name(monitorIndex);
 #endif
 #ifdef FLTK_USE_WAYLAND
             if (fl_wl_display())
-                out = getWaylandMonitorName(monitorIndex);
+                out = getWaylandName(monitorIndex);
 #endif
             return out;
         }